feat: add ROI stats, image sequence nav, overlay blend, scatter plot
- ROI (Ctrl+E): 드래그로 영역 선택, A/B/Diff 통계 창, 상태바 표시
- 시퀀스 탐색 (N/Shift+N): 디렉토리 내 이미지 natural sort, 현재/전체 표시
- Overlay/Blend 모드 (O): Blend/Curtain 두 방식, alpha 슬라이더, 전용 셰이더
- Scatter Plot (Ctrl+T): 채널별 A vs B 픽셀 분포 시각화
- AppImage: arm64/aarch64 아키텍처 자동 감지 지원
- ESC: ROI/Overlay/Scatter 창도 닫힘 처리

+++ src/image_save.cpp
@@ -170,9 +170,9 @@ static std::vector<uint8_t> compute_diff_cpu(const ImageEntry& imgA,
             get_pixel_f32(imgA, x, y, ar, ag, ab);
             get_pixel_f32(imgB, x, y, br, bg, bb);
 
-            float dr = std::fabsf(ar - br);
-            float dg = std::fabsf(ag - bg);
-            float db = std::fabsf(ab - bb);
+            float dr = std::fabs(ar - br);
+            float dg = std::fabs(ag - bg);
+            float db = std::fabs(ab - bb);
 
             float or_, og, ob;
